wifi-heatmap/requirements.txt
speedtest-cli
numpy
matplotlib
scipy
